deal with normalized interfaces
Kubernetes API validation does not allow to use some valid interface
names as Device names, so we need to normalize the name and we use an
attribute to export the real name.

The driver keeps a local cache for the exported devices so can sneek
into the cache to get the ifname if needed.

Change-Id: I7cbf0751a3470ba9891ca8a9bc8e8945a8d0845b